New curlscripts completed
Now create, show, and index all work.
I found that I had to authenticate views. When I removed requireToken,
I was able to see these in the API server. This allowed me to see INDEX
and show working properly.

I had to update Show to include the id and a slash in the url path.